Refactor helm-net into protocol and full features
- Split helm-net into feature-gated modules: protocol, config, and full
  node
- Move node logic into helm-net/src/node/
- Make protocol types available with the "protocol" feature
- Make network dependencies optional and feature-gated
- Update dependencies and features in Cargo.toml
- Update helm-cli and main crate to use new features and imports
- Refactor Input construction to use builder pattern
