+++ helm-cli/Cargo.toml
@@ -6,6 +6,8 @@ edition = "2024"
 [dependencies]
 # Local workspace crates
 helm-core = { path = "../helm-core" }
+helm-net = { path = "../helm-net", features = ["protocol"] }
+
 const-hex = "1.17.0"
 clap = { version = "4.5", features = ["derive"] }
 serde_json = "1.0"
