+++ helm-net/Cargo.toml
@@ -5,11 +5,20 @@ edition = "2024"
 
 [dependencies]
 helm-core = { path = "../helm-core" }
-libp2p = { version = "0.56", features = ["gossipsub", "mdns", "tcp", "noise", "yamux", "tokio", "request-response", "cbor"] }
-tokio = { version = "1.0", features = ["sync", "time", "macros"] }
-serde = { version = "1.0", features = ["derive"] }
-const-hex = "1.17.0"
-postcard = { version = "1.1", features = ["alloc"] }
-dotenv = "0.15.0"
-tracing = "0.1"
-ethnum = { version = "1.5.2", features = ["serde"] }
+
+serde = { version = "1.0", features = ["derive"], optional = true }
+const-hex = { version = "1.17.0", optional = true }
+ethnum = { version = "1.5.2", features = ["serde"], optional = true }
+
+libp2p = { version = "0.56", features = ["gossipsub", "mdns", "tcp", "noise", "yamux", "tokio", "request-response", "cbor"], optional = true }
+tokio = { version = "1.0", features = ["sync", "time", "macros"], optional = true }
+postcard = { version = "1.1", features = ["alloc"], optional = true }
+tracing = { version = "0.1", optional = true }
+
+dotenv = { version = "0.15.0", optional = true }
+
+[features]
+default = ["full"]
+protocol = ["serde", "const-hex", "ethnum"]
+full = ["libp2p", "tokio", "postcard", "tracing", "config", "protocol"]
+config = ["dotenv"]
