+++ helm-net/src/node/behavior.rs
